Subiendo los cambios de los scripts y ajuste de la tabla users

+++ Script-sp-surveys.sql
@@ -24,7 +24,8 @@ BEGIN
         en_id, 
         en_descripcion_pregunta, 
         tbl_usuarios_usu_id
-    FROM tbl_encuesta;
+    FROM tbl_encuesta
+    ORDER BY en_id DESC;
 END//
 DELIMITER ;
 
@@ -39,8 +40,9 @@ BEGIN
         CONCAT(u.usu_nombre, ' ', u.usu_apellido) AS nombre_completo
     FROM tbl_encuesta e
     -- Unir con la tabla tbl_usuarios para obtener el nombre completo
-    JOIN tbl_usuarios u ON e.tbl_usuarios_usu_id = u.usu_id;
-END//
+    JOIN tbl_usuarios u ON e.tbl_usuarios_usu_id = u.usu_id
+    ORDER BY en_id DESC;
+    END//
 DELIMITER ;
 
 -- Actualizar una encuesta
@@ -61,15 +63,32 @@ BEGIN
 END//
 DELIMITER ;
 
--- Eliminar una encuesta
+
+
+-- -- Eliminar una encuesta
+
 DELIMITER //
-CREATE PROCEDURE procDeleteSurvey(
-    IN v_en_id INT 
+CREATE PROCEDURE procDeleteSurvey (
+    IN v_en_id INT
 )
-BEGIN 
-    -- Eliminar la encuesta por su ID
-    DELETE FROM tbl_encuesta 
-    WHERE 
-        en_id = v_en_id;
-END//
+BEGIN
+    DECLARE v_exists INT DEFAULT 0;
+    -- Verificar si la encuesta existe
+    SELECT COUNT(*) INTO v_exists
+    FROM tbl_encuesta
+    WHERE en_id = v_en_id;
+
+    IF v_exists = 0 THEN
+        SIGNAL SQLSTATE '45000'
+        SET MESSAGE_TEXT = 'La encuesta no existe';
+    ELSE
+        -- Eliminar las respuestas relacionadas
+        DELETE FROM tbl_respuestas
+        WHERE tbl_encuesta_en_id = v_en_id;
+        -- Eliminar la encuesta
+        DELETE FROM tbl_encuesta
+        WHERE en_id = v_en_id;
+    END IF;
+END //
 DELIMITER ;
+
